refactor(daemon-bootstrap): convert to lib-only, add SIGTERM + PID file
